+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-12-08
+
 ### Added
 
 - Consider cluster deletion when `keep-until` label is set.
@@ -24,5 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/cluster-cleaner/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-cleaner/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/giantswarm/cluster-cleaner/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/giantswarm/cluster-cleaner/releases/tag/v0.1.0